requirements.txt
aiogram
python-dotenv
aiohttp-socks
